Move button definitions to calculator.cpp, move calculator.cpp/.h to src/ folder.
