+++ src/calculator.h
@@ -86,4 +86,16 @@ private:
     Button *digitButtons[NumDigitButtons];
 };
 
+#include <QToolButton>
+
+class Button : public QToolButton
+{
+    Q_OBJECT
+
+public:
+    explicit Button(const QString &text, QWidget *parent = 0);
+
+    QSize sizeHint() const Q_DECL_OVERRIDE;
+};
+
 #endif
